🩹(vm/utils/ssh): Copy SSH keys to shared folder
- Added a condition to check if the SSH key exists in the home directory of the vagrant user and if it does not exist in the shared folder.
- The SSH keys are now copied to the shared folder `/vagrant/.vagrant/ssh/`.

+++ vm/utils/ssh.sh
@@ -32,7 +32,7 @@ then
     fi
 
     # Copy the keys to the shared folder `/vagrant/.vagrant/ssh/`
-    if [-f /home/vagrant/.ssh/id_ed25519 ]
+    if [[ -f /home/vagrant/.ssh/id_ed25519 && ! -f /vagrant/.vagrant/ssh/id_ed25519 ]]
     then
         cp /home/vagrant/.ssh/id_* /vagrant/.vagrant/ssh/
     fi
